refactor camera and asset field
