+++ web/src/beta/ui/fields/AssetField.tsx
@@ -5,7 +5,7 @@ import { AcceptedFileFormat } from "@reearth/beta/features/Assets/types";
 import AssetModal from "@reearth/beta/features/Modals/AssetModal";
 import LayerStyleModal from "@reearth/beta/features/Modals/LayerStyleModal";
 import useFileUploaderHook from "@reearth/beta/hooks/useAssetUploader/hooks";
-import { TextInput, TextInputProps, Button, ButtonProps } from "@reearth/beta/lib/reearth-ui";
+import { TextInput, Button } from "@reearth/beta/lib/reearth-ui";
 import { checkIfFileType } from "@reearth/beta/utils/util";
 import { useT } from "@reearth/services/i18n";
 import { useNotification, useWorkspace } from "@reearth/services/state";
@@ -13,34 +13,26 @@ import { styled } from "@reearth/services/theme";
 
 import CommonField, { CommonFieldProps } from "./CommonField";
 
-export type AssetFieldProps = CommonFieldProps &
-  ButtonProps &
-  TextInputProps & {
-    entityType?: "image" | "file" | "layerStyle";
-    fileType?: "asset" | "URL" | "layerStyle";
-    fileFormat?: AcceptedFileFormat;
-    sceneId?: string;
-    onChange?: (value: string | undefined, name: string | undefined) => void;
-    maxWidth?: number;
-  };
+export type AssetFieldProps = CommonFieldProps & {
+  value?: string;
+  fileType?: "asset" | "URL" | "layerStyle";
+  entityType?: "image" | "file" | "layerStyle";
+  fileFormat?: AcceptedFileFormat;
+  sceneId?: string;
+  placeholder?: string;
+  onChange?: (value: string | undefined, name: string | undefined) => void;
+};
 
 const AssetField: FC<AssetFieldProps> = ({
   commonTitle,
   description,
-  appearance,
-  disabled,
-  iconButton,
   value,
-  placeholder,
-  onChange,
-  extendWidth,
   entityType,
   fileType,
   fileFormat,
   sceneId,
-  maxWidth,
-  size,
-  minWidth,
+  placeholder,
+  onChange,
 }) => {
   const t = useT();
   const [open, setOpen] = useState(false);
@@ -86,34 +78,21 @@ const AssetField: FC<AssetFieldProps> = ({
 
   return (
     <CommonField commonTitle={commonTitle} description={description}>
-      <AssetWrapper maxWidth={maxWidth}>
+      <AssetWrapper>
         <TextInput
           value={currentValue}
           onChange={handleChange}
-          extendWidth={extendWidth}
           placeholder={placeholder ?? t("Not set")}
         />
         {fileType === "asset" && (
           <ButtonWrapper>
-            <SelectionButton
-              appearance={appearance}
-              icon={"image"}
-              size={size}
-              disabled={disabled}
-              iconButton={iconButton}
-              title="Choose"
-              onClick={handleClick}
-              minWidth={minWidth}
-            />
-            <SelectionButton
-              appearance={appearance}
+            <Button icon={"image"} size="small" title="Choose" onClick={handleClick} extendWidth />
+            <Button
               icon={"uploadSimple"}
-              size={size}
-              disabled={disabled}
-              iconButton={iconButton}
+              size="small"
               title="Upload"
               onClick={handleFileUpload}
-              minWidth={minWidth}
+              extendWidth
             />
           </ButtonWrapper>
         )}
@@ -143,12 +122,9 @@ const AssetField: FC<AssetFieldProps> = ({
 
 export default AssetField;
 
-const AssetWrapper = styled("div")<{
-  maxWidth?: number;
-}>(({ theme, maxWidth }) => ({
+const AssetWrapper = styled("div")<{}>(({ theme }) => ({
   display: "flex",
   flexDirection: "column",
-  maxWidth: `${maxWidth}px`,
   gap: `${theme.spacing.smallest}px`,
   width: "100%",
 }));
@@ -158,7 +134,3 @@ const ButtonWrapper = styled("div")(({ theme }) => ({
   flexDirection: "row",
   gap: `${theme.spacing.smallest}px`,
 }));
-
-const SelectionButton = styled(Button)`
-  cursor: pointer;
-`;
